refactor: move form base and widgets to shared components
