+++ bciers/apps/registration/tests/components/form/FormBase.test.tsx
@@ -1,8 +1,8 @@
 import { fireEvent, render, screen } from "@testing-library/react";
 import { describe, expect } from "vitest";
 import React from "react";
-import FormBase from "@/app/components/form/FormBase";
-import readOnlyTheme from "@/app/components/form/readOnlyTheme";
+import FormBase from "@components/form/FormBase";
+import readOnlyTheme from "@components/form/theme/readOnlyTheme";
 
 const testSchema = {
   type: "object",
