+++ bciers/apps/registration/tests/components/form/widgets/EmailWidget.test.tsx
@@ -1,7 +1,7 @@
 import { userEvent } from "@testing-library/user-event";
 import { render, screen } from "@testing-library/react";
 import { RJSFSchema } from "@rjsf/utils";
-import FormBase from "@/app/components/form/FormBase";
+import FormBase from "@components/form/FormBase";
 import {
   checkNoValidationErrorIsTriggered,
   checkTextWidgetValidationStyles,
