+++ bciers/apps/registration/tests/components/form/widgets/OperatorSearchWidget.test.tsx
@@ -7,7 +7,7 @@ import {
   waitFor,
 } from "@testing-library/react";
 import { RJSFSchema } from "@rjsf/utils";
-import FormBase from "@/app/components/form/FormBase";
+import FormBase from "@components/form/FormBase";
 import { actionHandler } from "@/tests/setup/mocks";
 import {
   checkComboBoxWidgetValidationStyles,
