+++ bciers/apps/registration/tests/components/form/widgets/readOnly/ReadOnlyBooleanWidget.test.tsx
@@ -1,6 +1,6 @@
 import { render } from "@testing-library/react";
 import { RJSFSchema } from "@rjsf/utils";
-import FormBase from "@/app/components/form/FormBase";
+import FormBase from "@components/form/FormBase";
 
 const booleanFieldSchema = {
   type: "object",
