+++ bciers/libs/shared/components/src/lib/form/SingleStepTaskListForm.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from "@testing-library/react";
+
+import SingleStepTaskListForm from "./SingleStepTaskListForm";
+
+describe("SingleStepTaskListForm", () => {
+  it("should render successfully", () => {
+    const { baseElement } = render(<SingleStepTaskListForm />);
+    expect(baseElement).toBeTruthy();
+  });
+});
