+++ bciers/libs/shared/components/src/lib/form/widgets/OperatorSearchWidget.tsx
