+++ bciers/tsconfig.base.json
@@ -16,11 +16,12 @@
     "target": "es2015",
     "types": ["vitest/globals"],
     "paths": {
-      "@bciers/components/server": ["libs/shared/components/src/server.ts"],
+      "@/*": ["apps/registration/*"],
+      "@components/*": ["libs/shared/components/src/lib/*"],
       "@bciers/components": ["libs/shared/components/src/index.ts"],
+      "@bciers/components/server": ["libs/shared/components/src/server.ts"],
       "@bciers/img/*": ["libs/shared/img/*"],
-      "@bciers/styles": ["libs/shared/styles/src/index.ts"],
-      "@/*": ["apps/registration/*"]
+      "@bciers/styles": ["libs/shared/styles/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
